requirements.txt
streamlit==1.33.0
pandas>=2.2.2
transformers>=4.41.0
torch>=2.7.1
python-docx==0.8.11
PyPDF2==3.0.1
nltk==3.8.1
plotly==5.21.0
sentence-transformers==2.6.1
